CLOUDIFY-857 Added a test.
The test changes the default byon cloud DSL with one that has 3 machine templates, and assigns different services to different templates. The test then assert that services are on the correct machines, and checks uninstall and teardown. 

In two of the three templates, machine names are entered instead of their IPs in the byon-cloud.groovy file.

git-svn-id: svn://pc-lab14/SVN/cloudify/trunk/quality/frameworks/SGTest@120527 eb64e737-3616-4df0-8941-5ee2ae88103d